Move block related generators to models submodule
