+++ block-storage/src/test/scala/coop/rchain/blockstorage/BlockDagStorageTest.scala
@@ -5,13 +5,13 @@ import java.nio.file.StandardOpenOption
 import cats.implicits._
 import coop.rchain.shared.PathOps._
 import coop.rchain.catscontrib.TaskContrib.TaskOps
-import BlockGen._
 import cats.effect.Sync
 import coop.rchain.blockstorage.BlockDagRepresentation.Validator
 import coop.rchain.blockstorage.BlockStore.BlockHash
 import coop.rchain.blockstorage.util.byteOps._
 import coop.rchain.casper.protocol.BlockMessage
 import coop.rchain.metrics.Metrics.MetricsNOP
+import coop.rchain.models.blockImplicits._
 import coop.rchain.rspace.Context
 import coop.rchain.shared
 import coop.rchain.shared.Log
